requirements.txt
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SocketIO==4.2.1
Flask-SQLAlchemy==2.4.3
marshmallow==3.5.0
psycopg2-binary==2.8.5
pynt==0.8.2
pytest==6.0.1
pytest-cov==2.10.0
pytest-mock==3.2.0
requests==2.23.0
Scrapy==2.3.0
SQLAlchemy-Utils==0.36.6
uWSGI==2.0.18